+++ JotunnLib/Managers/PieceManager.cs
@@ -48,6 +48,7 @@ namespace Jotunn.Managers
 
         private const float PieceCategorySize = 700f;
         private const float PieceCategoryTabSize = 150f;
+        private const float PieceCategoryTabOffset = 50f;
 
         /// <summary>
         ///     Creates the piece table container and registers all hooks.
@@ -162,6 +163,7 @@ namespace Jotunn.Managers
         public Piece.PieceCategory AddPieceCategory(string table, string name)
         {
             Piece.PieceCategory categoryID;
+            bool isNew = false;
 
             // Get or create global category ID
             if (Enum.IsDefined(typeof(Piece.PieceCategory), name))
@@ -177,6 +179,7 @@ namespace Jotunn.Managers
                 categoryID = PieceCategories.Count + Piece.PieceCategory.Max;
                 PieceCategories.Add(name, categoryID);
                 PieceCategoryMax++;
+                isNew = true;
             }
 
             // Add category to table map
@@ -189,9 +192,26 @@ namespace Jotunn.Managers
                 PieceTableCategoriesMap[table].Add(name, categoryID);
             }
 
+            // When new categories are inserted in-game, directly create and update categories
+            if (isNew)
+            {
+                if (SceneManager.GetActiveScene().name == "main")
+                {
+                    CreatePieceTableCategories();
+                }
+                if (Hud.instance != null)
+                {
+                    CreateCategoryTabs();
+                }
+                if (Player.m_localPlayer != null)
+                {
+                    UpdatePieceCategories();
+                }
+            }
+
             return categoryID;
         }
-
+        
         /// <summary>
         ///     Get a list of all custom piece category names
         /// </summary>
@@ -316,7 +336,7 @@ namespace Jotunn.Managers
                 return;
             }
 
-            Logger.LogInfo($"Adding {PieceCategories.Count} custom piece table categories");
+            //Logger.LogInfo($"Adding {PieceCategories.Count} custom piece table categories");
 
             // All piece tables using categories
             foreach (var table in PieceTableMap.Values.Where(x => x.m_useCategories))
@@ -385,24 +405,27 @@ namespace Jotunn.Managers
                 }
             }
 
-            // Hook sceneLoaded for generation of custom category tabs
-            SceneManager.sceneLoaded += CreateCategoryTabs;
+            // Hook for generation of custom category tabs
+            On.Hud.Awake -= CreateCategoryTabs;
+            On.Hud.Awake += CreateCategoryTabs;
 
             // Hook for custom category tab toggle
+            On.Hud.UpdateBuild -= TogglePieceCategories;
             On.Hud.UpdateBuild += TogglePieceCategories;
         }
-
+        
         /// <summary>
         ///     Create tabs for the ingame GUI for every piece table category.
-        ///     Only executes when custom piece table categories were added.
+        ///     Only executes when custom piece table catego-ries were added.
         /// </summary>
-        private void CreateCategoryTabs(Scene scene, LoadSceneMode loadSceneMode)
+        private void CreateCategoryTabs(On.Hud.orig_Awake orig, Hud self)
         {
-            if (scene.name != "main")
-            {
-                return;
-            }
+            orig(self);
+            CreateCategoryTabs();
+        }
 
+        private void CreateCategoryTabs()
+        {
             // Get the GUI elements
             GameObject root = Hud.instance.m_pieceCategoryRoot;
             if (root.GetComponent<RectMask2D>() == null)
@@ -413,6 +436,7 @@ namespace Jotunn.Managers
                 Transform border = root.transform.Find("TabBorder");
                 border?.SetParent(root.transform.parent, true);
             }
+
             List<string> newNames = new List<string>(Hud.instance.m_buildCategoryNames);
             List<GameObject> newTabs = new List<GameObject>(Hud.instance.m_pieceCategoryTabs);
 
@@ -435,24 +459,27 @@ namespace Jotunn.Managers
                     newTabs.Add(newTab);
                 }
             }
-
-            // Reorder tabs
-            float offset = 0f;
-            foreach (GameObject go in newTabs)
-            {
-                go.SetMiddleLeft();
-                go.SetWidth(PieceCategoryTabSize);
-                RectTransform tf = go.transform as RectTransform;
-                tf.anchoredPosition = new Vector2(offset, 0f);
-                offset += PieceCategoryTabSize;
-            }
-
+            
             // Replace the HUD arrays
             Hud.instance.m_buildCategoryNames = newNames.ToList();
             Hud.instance.m_pieceCategoryTabs = newTabs.ToArray();
+        }
 
-            // Remove ourself from the SceneManager
-            SceneManager.sceneLoaded -= CreateCategoryTabs;
+        /// <summary>
+        ///     Reorder piece table tabs if the table if opened currently
+        /// </summary>
+        private void UpdatePieceCategories()
+        {
+            if (!Player.m_localPlayer)
+            {
+                return;
+            }
+
+            var table = Player.m_localPlayer.m_buildPieces;
+            if (table != null && table.m_useCategories && PieceTableCategoriesMap.TryGetValue(table.name, out var categories))
+            {
+                categories.ReorderTableTabs();
+            }
         }
 
         /// <summary>
@@ -464,9 +491,9 @@ namespace Jotunn.Managers
             try
             {
                 var table = Player.m_localPlayer.m_buildPieces;
-                if (table != null && table.m_useCategories && PieceTableCategoriesMap.ContainsKey(table.name))
+                if (table != null && table.m_useCategories && PieceTableCategoriesMap.TryGetValue(table.name, out var categories))
                 {
-                    PieceTableCategoriesMap[table.name].Toggle(Hud.instance.m_pieceSelectionWindow.activeSelf);
+                    categories.Toggle(Hud.instance.m_pieceSelectionWindow.activeSelf);
                 }
             }
             catch (Exception ex)
@@ -693,6 +720,21 @@ namespace Jotunn.Managers
                 }
             }
 
+            internal void ReorderTableTabs()
+            {
+                if (currentActive != null && currentActive == this)
+                {
+                    // Activate all tabs for this categories
+                    foreach (GameObject tab in Hud.instance.m_pieceCategoryTabs)
+                    {
+                        tab.SetActive(Keys.Contains(tab.name));
+                    }
+
+                    // Reorder tabs
+                    ReorderActiveTabs();
+                }
+            }
+
             private void ReorderActiveTabs()
             {
                 float offset = 0f;
@@ -700,6 +742,7 @@ namespace Jotunn.Managers
                 {
                     go.SetMiddleLeft();
                     go.SetHeight(30f);
+                    go.SetWidth(PieceCategoryTabSize);
                     RectTransform tf = (go.transform as RectTransform);
                     tf.anchoredPosition = new Vector2(offset, 0f);
                     offset += PieceCategoryTabSize;
@@ -783,22 +826,22 @@ namespace Jotunn.Managers
                 var tab = Hud.instance.m_pieceCategoryTabs[(int)selectedCategory];
                 lastCategory = selectedCategory;
 
-                float minX = (tab.transform as RectTransform).anchoredPosition.x;
+                float minX = tab.GetComponent<RectTransform>().anchoredPosition.x - PieceCategoryTabOffset;
                 if (minX < 0f)
                 {
-                    float offsetX = selectedCategory == 0 ? minX * -1 : PieceCategoryTabSize;
+                    float offsetX = selectedCategory == Values.Min() ? minX * -1 - PieceCategoryTabOffset : minX * -1;
                     foreach (GameObject go in Hud.instance.m_pieceCategoryTabs)
                     {
-                        (go.transform as RectTransform).anchoredPosition += new Vector2(offsetX, 0);
+                        go.GetComponent<RectTransform>().anchoredPosition += new Vector2(offsetX, 0);
                     }
                 }
-                float maxX = (tab.transform as RectTransform).anchoredPosition.x + PieceCategoryTabSize;
+                float maxX = tab.GetComponent<RectTransform>().anchoredPosition.x + PieceCategoryTabSize + PieceCategoryTabOffset;
                 if (maxX > PieceCategorySize)
                 {
-                    float offsetX = selectedCategory == Values.Max() ? maxX - PieceCategorySize : PieceCategoryTabSize;
+                    float offsetX = selectedCategory == Values.Max() ? maxX - PieceCategorySize - PieceCategoryTabOffset: maxX - PieceCategorySize;
                     foreach (GameObject go in Hud.instance.m_pieceCategoryTabs)
                     {
-                        (go.transform as RectTransform).anchoredPosition -= new Vector2(offsetX, 0);
+                        go.GetComponent<RectTransform>().anchoredPosition -= new Vector2(offsetX, 0);
                     }
                 }
             }