+++ JotunnLib/Managers/SynchronizationManager.cs
@@ -548,17 +548,7 @@ namespace Jotunn.Managers
                     }
 
                     // Set buttons if changed
-                    string buttonName = cx.GetBoundButtonName();
-                    if (cx.SettingType == typeof(KeyCode) && ZInput.instance != null &&
-                        ZInput.instance.m_buttons.ContainsKey(buttonName))
-                    {
-                        ZInput.instance.Setbutton(buttonName, (KeyCode)cx.BoxedValue);
-                    }
-                    if (cx.SettingType == typeof(KeyboardShortcut) && ZInput.instance != null &&
-                        ZInput.instance.m_buttons.ContainsKey(buttonName))
-                    {
-                        ZInput.instance.Setbutton(buttonName, ((KeyboardShortcut)cx.BoxedValue).MainKey);
-                    }
+                    SetInputButtons(cx);
                 }
             }
 
@@ -589,6 +579,53 @@ namespace Jotunn.Managers
                 
                 // Rebuild config cache
                 CacheConfigurationValues();
+            }
+        }
+
+        private void SetInputButtons(ConfigEntryBase entry)
+        {
+            if (ZInput.instance == null)
+            {
+                return;
+            }
+
+            string buttonName = entry.GetBoundButtonName();
+
+            if (string.IsNullOrEmpty(buttonName))
+            {
+                return;
+            }
+
+            ZInput.ButtonDef def;
+
+            if (entry.SettingType == typeof(KeyCode) &&
+                ZInput.instance.m_buttons.TryGetValue(buttonName, out def))
+            {
+                def.m_key = (KeyCode)entry.BoxedValue;
+            }
+
+            if (entry.SettingType == typeof(KeyboardShortcut) &&
+                ZInput.instance.m_buttons.TryGetValue(buttonName, out def))
+            {
+                def.m_key = ((KeyboardShortcut)entry.BoxedValue).MainKey;
+            }
+
+            if (entry.SettingType == typeof(InputManager.GamepadButton) &&
+                ZInput.instance.m_buttons.TryGetValue($"Joy!{buttonName}", out def))
+            {
+                var keyCode = InputManager.GetGamepadKeyCode((InputManager.GamepadButton)entry.BoxedValue);
+                var axis = InputManager.GetGamepadAxis((InputManager.GamepadButton)entry.BoxedValue);
+                
+                if (!string.IsNullOrEmpty(axis))
+                {
+                    bool invert = axis.StartsWith("-");
+                    def.m_axis = axis.TrimStart('-');
+                    def.m_inverted = invert;
+                }
+                else
+                {
+                    def.m_key = keyCode;
+                }
             }
         }
 
@@ -804,31 +841,21 @@ namespace Jotunn.Managers
                             entry.BoxedValue = TomlTypeConverter.ConvertToValue(serializedValue, entry.SettingType);
 
                             // Set buttons after receive
-                            string buttonName = entry.GetBoundButtonName();
-                            if (entry.SettingType == typeof(KeyCode) && ZInput.instance != null &&
-                                ZInput.instance.m_buttons.ContainsKey(buttonName))
-                            {
-                                ZInput.instance.Setbutton(buttonName, (KeyCode)entry.BoxedValue);
-                            }
-                            if (entry.SettingType == typeof(KeyboardShortcut) && ZInput.instance != null &&
-                                ZInput.instance.m_buttons.ContainsKey(buttonName))
-                            {
-                                ZInput.instance.Setbutton(buttonName, ((KeyboardShortcut)entry.BoxedValue).MainKey);
-                            }
+                            SetInputButtons(entry);
                         }
                         else
                         {
-                            Logger.LogError($"Setting for GUID: {modguid}, Section {section}, Key {key} is not syncable");
+                            Logger.LogWarning($"Setting for GUID: {modguid}, Section {section}, Key {key} is not syncable");
                         }
                     }
                     else
                     {
-                        Logger.LogError($"Did not find Value for GUID: {modguid}, Section {section}, Key {key}");
+                        Logger.LogWarning($"Did not find Value for GUID: {modguid}, Section {section}, Key {key}");
                     }
                 }
                 else
                 {
-                    Logger.LogError($"No plugin with GUID {modguid} is loaded");
+                    Logger.LogWarning($"No plugin with GUID {modguid} is loaded");
                 }
 
                 numberOfEntries--;