+++ CHANGELOG.md
@@ -1,9 +1,33 @@
 ﻿# Changelog
 
+## Version 2.3.12
+* Added RenderManager.Render() which renders the given GameObject in the same frame instead of waiting for the next frame. Marked EnqueueRender obsolete.
+* Force unload mod's loaded asset bundles on Game.OnApplicationQuit to prevent the Unity engine from crashing
+* Depend on BepInExPack-5.4.1601 for the new Unity engine corlibs
+
+## Version 2.3.10
+* Added custom piece table category injection at runtime
+* Added RenderRequest for the RenderManager to define options for the render process
+* Fixed mock resolving of generic List types
+* Fixed CustomGUI anchor settings
+* Fixed Dropdown list sizing (thx joeyparrish)
+
+## Version 2.3.9
+* Added the possibility to define vanilla console command modifiers in ConsoleCommand (thx joeyparrish)
+* Added the possibility to define command options in ConsoleCommand
+* Fixed exception for mods loaded without PluginInfo.Location set
+* Fixed NRE on missing KeyHint objects
+
 ## Version 2.3.8
+* Added RenderManager to render Sprites from GameObjects at runtime (thx MSchmoecker)
+* Added GamepadButton to the InputManager and ButtonConfig - custom inputs can now define a gamepad button corresponding to the keyboard input
+* Added gamepad buttons to KeyHints for custom inputs as well as vanilla key overrides
+* Gamepad buttons can be defined in the mod settings if they are bound to a config
+* Refactored custom KeyHints into their own KeyHintManager and obsoleted the API in the GUIManager
 * Fixed automatic mod recognition from filesystem paths (thx Digitalroot)
 * Fixed duplication check on ObjectDB not always working correctly
 * Fixed NRE in ModCompat for VersionCheckOnly mods
+* Fixed a strange hard crash when using GUIManager.IsHeadless()
 
 ## Version 2.3.7
 * Fixed translation of the custom skill raise message